requirements.txt
Flask>=0.9
Flask-SQLAlchemy
gunicorn>=0.16.1
# patched because crash on PATCH
-e git+http://github.com/bjourne/flask-restless.git#egg=restless
# httplib2 misses certs
-e git+http://github.com/bjourne/python-oauth2.git@disable-ssl-check#egg=python-oauth
# Because unicode strings with encodings
-e git+http://github.com/bjourne/flask-oauth.git#egg=flask-oauth
psycopg2
Flask-Login
requests>=0.14.1
lxml>=2.3.2
PIL>=1.1.7
unidecode>=0.04.5
